requirements.txt

python-dotenv==1.0.1
pandas==2.2.2
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
plotly==5.24.1
streamlit==1.37.1
pyodbc==5.1.0
pydantic==2.8.2
